[DE-2259]: Added Ultralytics instructions for the neutron converter. Using YOLOv8s for the examples. Also added instructions for quantization.
